reduce boilerplate
