--- components/footer/elements/freshworks-dark.tsx
@@ -1,141 +1,0 @@
-import { StyleConfig } from '@chakra-ui/theme-tools';
-import { FC, SVGProps } from 'react';
-
-type Props = StyleConfig & SVGProps<SVGSVGElement>;
-
-export const FreshworksDark: FC<Props> = () => {
-  return (
-    <svg
-      width="138"
-      height="29"
-      viewBox="0 0 138 29"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clip-path="url(#clip0_122_428)">
-        <path
-          d="M33.6269 0.779877C30.1539 0.779877 28.9349 3.20303 28.9349 5.46463V6.84929H28.1988C27.6699 6.84929 27.2559 7.26468 27.2559 7.7955C27.2559 8.32627 27.6699 8.74167 28.1988 8.74167H28.9349V18.2958C28.9349 18.8497 29.3718 19.3112 29.9469 19.3112C30.4989 19.3112 30.9589 18.8728 30.9589 18.2958V8.74167H33.1899C33.7189 8.74167 34.1329 8.32627 34.1329 7.7955C34.1329 7.26468 33.7189 6.84929 33.1899 6.84929H30.9589V5.46463C30.9589 4.19539 31.4189 2.69532 33.6499 2.69532C34.1789 2.69532 34.5929 2.27992 34.5929 1.74914C34.5699 1.19527 34.1559 0.779877 33.6269 0.779877Z"
-          fill="white"
-        />
-        <path
-          d="M41.4462 6.52655C39.9515 6.52655 38.6862 7.14965 37.9734 8.16509V7.63428C37.9734 7.08041 37.5363 6.61887 36.9614 6.61887C36.4093 6.61887 35.9493 7.05732 35.9493 7.63428V18.2962C35.9493 18.85 36.3864 19.3116 36.9614 19.3116C37.5134 19.3116 37.9734 18.8731 37.9734 18.2962V12.4113C37.9734 9.94205 39.2845 8.46507 41.4462 8.46507C41.9753 8.46507 42.4122 8.02658 42.4122 7.49581C42.4355 6.96504 41.9982 6.52655 41.4462 6.52655Z"
-          fill="white"
-        />
-        <path
-          d="M53.7512 11.7655C53.7512 11.1885 53.6591 9.68848 52.7852 8.44229C51.8884 7.14992 50.4851 6.48068 48.5764 6.48068C45.0343 6.48068 43.0101 8.8346 43.0101 12.9424C43.0101 17.1887 45.0343 19.4503 48.8291 19.4503C51.3823 19.4503 52.7852 18.1811 53.0383 17.9272C53.1992 17.7657 53.3831 17.5579 53.3831 17.2118C53.3831 16.7041 52.9691 16.2887 52.4633 16.2887C52.2103 16.2887 52.0493 16.3579 51.8884 16.4733C51.1521 17.0272 50.3012 17.581 48.8524 17.581C46.5752 17.581 45.1264 16.081 45.0572 13.6348H52.5321C53.7512 13.6117 53.7512 12.5039 53.7512 11.7655ZM48.5764 8.32688C50.6464 8.32688 51.7271 9.50384 51.7733 11.8116H45.0801C45.2644 9.57307 46.5064 8.32688 48.5764 8.32688Z"
-          fill="white"
-        />
-        <path
-          d="M60.5132 11.9269C59.0182 11.4193 57.7303 10.9577 57.7303 9.84996C57.7303 8.62686 59.0874 8.44221 59.8924 8.44221C61.3412 8.44221 62.0542 8.94993 62.6291 9.5038C62.8134 9.68844 63.0431 9.78073 63.2962 9.78073C63.8023 9.78073 64.2163 9.36533 64.2163 8.85765C64.2163 8.58072 64.1012 8.39607 64.0091 8.25761C63.1582 7.12679 61.7094 6.50369 59.8924 6.50369C57.3163 6.50369 55.7291 7.79606 55.7291 9.84996C55.7291 12.3654 57.9371 13.127 59.8691 13.7732C61.3412 14.2809 62.6062 14.6963 62.6062 15.7809C62.6062 17.2579 60.9731 17.4887 59.9842 17.4887C58.6504 17.4887 57.6152 17.0733 56.7643 16.2425C56.5112 15.9886 56.2352 15.9425 56.0743 15.9425C55.5682 15.9425 55.1542 16.3579 55.1542 16.8656C55.1542 17.0733 55.2234 17.2348 55.3843 17.4656C56.1431 18.4349 57.8683 19.4272 59.9842 19.4272C62.8822 19.4272 64.6074 18.0656 64.6074 15.804C64.6074 13.3578 62.4452 12.5962 60.5132 11.9269Z"
-          fill="white"
-        />
-        <path
-          d="M72.4504 6.48079C71.0704 6.48079 69.8284 7.03465 69.1384 7.91159V1.0806C69.1384 0.526737 68.7014 0.0651855 68.1265 0.0651855C67.5745 0.0651855 67.1143 0.503662 67.1143 1.0806V18.3196C67.1143 18.8735 67.5516 19.3351 68.1265 19.3351C68.6785 19.3351 69.1384 18.8966 69.1384 18.3196V11.8348C69.1384 9.50395 70.2653 8.32699 72.4504 8.32699C74.5663 8.32699 75.4405 9.29625 75.4405 11.6502V18.3196C75.4405 18.8735 75.8775 19.3351 76.4524 19.3351C77.0044 19.3351 77.4643 18.8966 77.4643 18.3196V11.304C77.4184 8.0962 75.7395 6.48079 72.4504 6.48079Z"
-          fill="white"
-        />
-        <path
-          d="M94.1856 6.64285C93.7483 6.64285 93.4035 6.89673 93.2425 7.31213L90.3675 15.5278L88.0215 7.31213C87.9294 6.98902 87.5845 6.64285 87.0784 6.64285C86.5723 6.64285 86.2275 6.98902 86.1354 7.31213L83.7894 15.5047L80.9143 7.28904C80.7534 6.87364 80.4085 6.64285 79.9716 6.64285C79.4196 6.64285 78.9594 7.08134 78.9594 7.65829C78.9594 7.79676 78.9823 7.93523 79.0515 8.07369L82.8925 18.6663C83.0535 19.0817 83.3983 19.3125 83.8356 19.3125C84.2955 19.3125 84.6865 19.0125 84.8016 18.5971L87.0555 10.8199L89.3094 18.5971C89.4244 19.0125 89.8384 19.3125 90.2754 19.3125C90.7123 19.3125 91.0575 19.0587 91.2184 18.6433L95.0595 8.0506C95.1283 7.91213 95.1516 7.75058 95.1516 7.61211C95.1745 7.08134 94.7376 6.64285 94.1856 6.64285Z"
-          fill="white"
-        />
-        <path
-          d="M101.499 6.48068C97.8424 6.48068 95.8412 8.78846 95.8412 12.9424C95.8412 17.1425 97.8424 19.4503 101.499 19.4503C105.156 19.4503 107.157 17.1425 107.157 12.9424C107.157 8.76537 105.133 6.48068 101.499 6.48068ZM101.499 17.6041C99.0614 17.6041 97.8653 16.081 97.8653 12.9655C97.8653 9.89618 99.0844 8.34997 101.499 8.34997C103.914 8.34997 105.133 9.89618 105.133 12.9655C105.133 16.081 103.937 17.6041 101.499 17.6041Z"
-          fill="white"
-        />
-        <path
-          d="M114.932 6.52655C113.437 6.52655 112.172 7.14965 111.459 8.16509V7.63428C111.459 7.08041 111.022 6.61887 110.447 6.61887C109.895 6.61887 109.435 7.05732 109.435 7.63428V18.2962C109.435 18.85 109.872 19.3116 110.447 19.3116C110.999 19.3116 111.459 18.8731 111.459 18.2962V12.4113C111.459 9.94205 112.77 8.46507 114.932 8.46507C115.461 8.46507 115.898 8.02658 115.898 7.49581C115.898 6.96504 115.461 6.52655 114.932 6.52655Z"
-          fill="white"
-        />
-        <path
-          d="M122.567 11.7886L126.731 8.44236C127.075 8.16543 127.121 7.81927 127.121 7.63466C127.121 7.0808 126.684 6.61926 126.109 6.61926C125.948 6.61926 125.672 6.64231 125.419 6.8731L119.508 11.7194V1.0806C119.508 0.526737 119.071 0.0651855 118.496 0.0651855C117.944 0.0651855 117.485 0.503662 117.485 1.0806V18.3196C117.485 18.8735 117.922 19.3351 118.496 19.3351C119.048 19.3351 119.508 18.8966 119.508 18.3196V14.1656L120.911 13.0348L125.442 18.8735C125.695 19.1735 125.971 19.3351 126.293 19.3351C126.845 19.3351 127.306 18.8966 127.306 18.3196C127.306 18.1581 127.283 17.9273 127.098 17.6965L122.567 11.7886Z"
-          fill="white"
-        />
-        <path
-          d="M133.906 11.9269C132.411 11.4193 131.123 10.9577 131.123 9.84996C131.123 8.62686 132.48 8.44221 133.285 8.44221C134.734 8.44221 135.447 8.94993 136.022 9.5038C136.206 9.68844 136.436 9.78073 136.689 9.78073C137.195 9.78073 137.609 9.36533 137.609 8.85765C137.609 8.55763 137.448 8.34993 137.402 8.25761C136.551 7.12679 135.102 6.50369 133.285 6.50369C130.709 6.50369 129.122 7.79606 129.122 9.84996C129.122 12.3654 131.33 13.127 133.262 13.7732C134.734 14.2809 135.999 14.6963 135.999 15.7809C135.999 17.2579 134.366 17.4887 133.377 17.4887C132.043 17.4887 131.008 17.0733 130.157 16.2425C129.904 15.9886 129.651 15.9425 129.467 15.9425C128.961 15.9425 128.547 16.3579 128.547 16.8656C128.547 17.0733 128.616 17.2348 128.777 17.4656C129.536 18.4349 131.261 19.4272 133.377 19.4272C136.275 19.4272 138 18.0656 138 15.804C138 13.3578 135.838 12.5962 133.906 11.9269Z"
-          fill="white"
-        />
-        <path
-          d="M0.138184 17.7431C0.529183 20.3047 1.72519 22.6125 3.45019 24.3433L7.47519 20.3047C7.47519 16.22 9.59121 12.2506 13.3632 10.0582L0.138184 17.7431Z"
-          fill="#9B65C3"
-        />
-        <path
-          d="M7.47519 20.3042C5.47419 20.3042 3.45019 19.7965 1.58719 18.7118C1.08119 18.4118 0.598183 18.0887 0.138184 17.7195C0.529183 20.2811 1.72519 22.5889 3.45019 24.3197L7.47519 20.3042Z"
-          fill="#9B65C3"
-        />
-        <path
-          d="M7.47459 20.3036L3.44958 24.3422C5.19759 26.0961 7.47459 27.2962 10.0276 27.6654L17.6636 14.3958C15.5016 18.1805 11.5456 20.3036 7.47459 20.3036Z"
-          fill="#BE63C5"
-        />
-        <path
-          d="M9.06159 26.235C7.9806 24.3657 7.47459 22.3348 7.47459 20.3271L3.44958 24.3657C5.19759 26.1196 7.47459 27.3196 10.0276 27.6889C9.68259 27.2273 9.3606 26.7427 9.06159 26.235Z"
-          fill="#BF63C6"
-        />
-        <path
-          d="M22.4713 4.79507L13.3862 10.0799C9.61421 12.2723 5.12921 12.1107 1.6102 10.0799C0.598196 11.8107 0.0231934 13.8415 0.0231934 15.9877C0.0231934 16.5878 0.0691933 17.1647 0.161193 17.7417C0.621193 18.1109 1.1042 18.434 1.6102 18.734C3.4732 19.8186 5.49723 20.3264 7.49823 20.3264L22.7243 5.04895C22.6322 4.95663 22.5632 4.86431 22.4713 4.79507Z"
-          fill="#3278B1"
-        />
-        <path
-          d="M1.58663 18.7352C3.44963 19.8198 5.47365 20.3276 7.47465 20.3276C7.47465 16.2428 9.59064 12.2734 13.3627 10.0811L0.114624 17.7429C0.597626 18.089 1.08063 18.4352 1.58663 18.7352Z"
-          fill="#3278B2"
-        />
-        <path
-          d="M1.587 10.0811C0.575001 11.8119 0 13.8427 0 15.9889C0 16.5889 0.0460003 17.1659 0.138 17.7429L13.386 10.0811C9.59104 12.2504 5.106 12.1119 1.587 10.0811Z"
-          fill="#45A4EC"
-        />
-        <path
-          d="M22.7239 5.0276C22.6549 4.95837 22.5629 4.86605 22.4709 4.79681C22.3559 4.70453 22.2409 4.61221 22.1028 4.54297C21.6658 4.28913 21.1599 4.15067 20.6308 4.15067H11.7988C7.42882 4.15067 3.6338 6.52767 1.58679 10.0586C5.1058 12.0894 9.5908 12.2509 13.3628 10.0586C9.5908 12.2509 7.47481 16.2203 7.47481 20.305C11.5458 20.305 15.5018 18.1819 17.6869 14.3972L22.9309 5.28148C22.8619 5.18916 22.7929 5.09684 22.7239 5.0276Z"
-          fill="#19BB7D"
-        />
-        <path
-          d="M13.3858 10.0805L22.4708 4.81879C22.3558 4.72647 22.2409 4.63414 22.1028 4.56491L1.58679 10.0805C5.1058 12.1113 9.5908 12.2498 13.3858 10.0805Z"
-          fill="#08C7FB"
-        />
-        <path
-          d="M20.6308 4.15067H11.7988C7.42882 4.15067 3.6338 6.52767 1.58679 10.0585L22.1028 4.54297C21.6658 4.31218 21.1828 4.15067 20.6308 4.15067Z"
-          fill="#59F2F6"
-        />
-        <path
-          d="M17.6873 14.3967L22.9313 5.28101C22.8623 5.18869 22.7933 5.11945 22.7012 5.02713L7.47522 20.3046C7.47522 22.3123 7.98123 24.3432 9.06222 26.2125C9.36123 26.7201 9.68322 27.2048 10.0512 27.6664C10.6262 27.7587 11.2012 27.8048 11.7992 27.8048C13.9382 27.8048 15.9623 27.2279 17.6873 26.2125C15.6632 22.6816 15.5022 18.1814 17.6873 14.3967Z"
-          fill="#DA3757"
-        />
-        <path
-          d="M7.47522 20.3036C7.47522 22.3114 7.98123 24.3422 9.06222 26.2115C9.36123 26.7192 9.68322 27.2039 10.0512 27.6654L17.6873 14.3958C15.5022 18.1805 11.5462 20.3036 7.47522 20.3036Z"
-          fill="#D33C4E"
-        />
-        <path
-          d="M17.6874 14.3958L10.0514 27.6885C10.6264 27.7808 11.2014 27.827 11.7994 27.827C13.9384 27.827 15.9624 27.25 17.6874 26.2346C15.6634 22.6806 15.5024 18.1805 17.6874 14.3958Z"
-          fill="#EE5A24"
-        />
-        <path
-          d="M23.1847 5.64896C23.1157 5.51049 23.0237 5.39512 22.9317 5.27974C22.8627 5.18742 22.7937 5.11819 22.7017 5.02586C22.6327 4.95663 22.5407 4.86431 22.4487 4.79507L13.3867 10.0799C9.61468 12.2723 7.49866 16.2416 7.49866 20.3264C11.5697 20.3264 15.5257 18.2032 17.7107 14.4185C15.5257 18.2032 15.6867 22.7034 17.7107 26.2343C21.2297 24.1804 23.5987 20.3725 23.5987 15.9878V7.12594C23.5757 6.57208 23.4377 6.06436 23.1847 5.64896Z"
-          fill="#8BDF55"
-        />
-        <path
-          d="M22.4717 4.79507L13.3867 10.0799C9.61468 12.2723 7.49866 16.2416 7.49866 20.3264L22.7247 5.04895C22.6327 4.95663 22.5637 4.86431 22.4717 4.79507Z"
-          fill="#25C16F"
-        />
-        <path
-          d="M7.47522 20.3045C11.5462 20.3045 15.5022 18.1814 17.6873 14.3967L22.9313 5.28101C22.8623 5.18869 22.7933 5.11945 22.7012 5.02713L7.47522 20.3045Z"
-          fill="#8BDF55"
-        />
-        <path
-          d="M22.9314 5.2811L17.6874 14.3968C15.5024 18.1815 15.6633 22.6817 17.6874 26.2125L23.1844 5.62726C23.1154 5.51189 23.0234 5.39651 22.9314 5.2811Z"
-          fill="#FFBB00"
-        />
-        <path
-          d="M17.6879 26.2348C21.2069 24.1809 23.5759 20.3731 23.5759 15.9883V7.12649C23.5759 6.59568 23.4379 6.08799 23.1849 5.64951L17.6879 26.2348Z"
-          fill="#FFA800"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_122_428">
-          <rect
-            width="138"
-            height="28"
-            fill="white"
-            transform="translate(0 0.0651855)"
-          />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
