--- components/header/elements/desktop-logo-dark.tsx
@@ -1,25 +1,0 @@
-import { StyleConfig } from '@chakra-ui/theme-tools';
-import { FC, SVGProps } from 'react';
-
-type Props = StyleConfig & SVGProps<SVGSVGElement>;
-
-export const DesktopLogoDark: FC<Props> = () => {
-  return (
-    <svg
-      width="104"
-      height="75"
-      viewBox="0 7 104 95"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M11.4489 44.1819L19.9062 71.8807H20.2202L28.6776 44.1819H38.9077L26.1477 82.0001H13.9787L1.21875 44.1819H11.4489ZM52.8356 82.5356C49.8687 82.5356 47.3081 81.9508 45.1538 80.7813C43.0117 79.5995 41.3621 77.9191 40.2049 75.7401C39.06 73.5488 38.4876 70.9451 38.4876 67.929C38.4876 64.9991 39.0662 62.4385 40.2234 60.2472C41.3806 58.0436 43.0117 56.3324 45.1168 55.1137C47.2219 53.8826 49.7025 53.2671 52.5586 53.2671C54.5775 53.2671 56.4241 53.581 58.0984 54.2089C59.7726 54.8367 61.2191 55.7662 62.4379 56.9972C63.6566 58.2283 64.6045 59.7486 65.2816 61.5583C65.9587 63.3556 66.2972 65.4177 66.2972 67.7444V69.9972H41.6452V64.7529H57.8952C57.8829 63.7927 57.6552 62.9371 57.212 62.1861C56.7688 61.4352 56.1594 60.8504 55.3839 60.4319C54.6206 60.001 53.7404 59.7856 52.7433 59.7856C51.7338 59.7856 50.829 60.0133 50.0288 60.4688C49.2286 60.912 48.5946 61.5214 48.1268 62.2969C47.659 63.0602 47.4128 63.9281 47.3881 64.9006V70.2373C47.3881 71.3945 47.6159 72.4101 48.0714 73.2842C48.5269 74.1459 49.1732 74.8168 50.0103 75.2969C50.8474 75.777 51.8446 76.0171 53.0018 76.0171C53.802 76.0171 54.5283 75.9063 55.1808 75.6847C55.8332 75.4631 56.3933 75.1369 56.8612 74.706C57.329 74.2752 57.6798 73.7458 57.9137 73.118L66.2049 73.358C65.8602 75.2169 65.1031 76.8358 63.9336 78.2145C62.7764 79.581 61.256 80.6459 59.3725 81.4092C57.489 82.1601 55.31 82.5356 52.8356 82.5356ZM79.9389 44.1819V82.0001H70.9091V44.1819H79.9389ZM101.696 53.6364V60.2842H83.8029V53.6364H101.696ZM87.5515 46.841H96.5813V73.081C96.5813 73.635 96.6675 74.0843 96.8398 74.429C97.0245 74.7614 97.2892 75.0015 97.6339 75.1492C97.9786 75.2846 98.391 75.3523 98.8711 75.3523C99.2158 75.3523 99.579 75.3216 99.9606 75.26C100.355 75.1861 100.65 75.1246 100.847 75.0753L102.213 81.5938C101.783 81.7169 101.173 81.8708 100.385 82.0555C99.6097 82.2401 98.6803 82.3571 97.5969 82.4063C95.4795 82.5048 93.6637 82.2586 92.1495 81.6677C90.6476 81.0645 89.4966 80.1288 88.6964 78.8609C87.9085 77.5929 87.5269 75.9986 87.5515 74.0782V46.841Z"
-        fill="white"
-      />
-      <path
-        d="M51 29.8352L87.7052 20.0001L90.0346 28.6934L53.3294 38.5285L51 29.8352Z"
-        fill="white"
-      />
-    </svg>
-  );
-};
