--- components/header/elements/logo-dark.tsx
@@ -1,38 +1,0 @@
-import { StyleConfig } from '@chakra-ui/theme-tools';
-import { FC, SVGProps } from 'react';
-
-type Props = StyleConfig & SVGProps<SVGSVGElement>;
-
-export const LogoDark: FC<Props> = () => {
-  return (
-    <svg
-      width="137"
-      height="39"
-      viewBox="15 0 99 21"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M50.3884 0.43219L54.8285 14.9741H54.9933L59.4334 0.43219H64.8043L58.1053 20.2867H51.7165L45.0176 0.43219H50.3884Z"
-        fill="white"
-      />
-      <path
-        d="M72.1164 20.5679C70.5588 20.5679 69.2145 20.2609 68.0835 19.6469C66.9589 19.0264 66.0928 18.1442 65.4853 17.0003C64.8842 15.8498 64.5837 14.4829 64.5837 12.8995C64.5837 11.3612 64.8875 10.0169 65.495 8.8665C66.1025 7.70961 66.9589 6.81124 68.0641 6.17139C69.1692 5.52509 70.4716 5.20194 71.971 5.20194C73.0309 5.20194 74.0004 5.36674 74.8794 5.69636C75.7583 6.02598 76.5178 6.51394 77.1576 7.16024C77.7974 7.80655 78.2951 8.60474 78.6506 9.55481C79.006 10.4984 79.1838 11.581 79.1838 12.8025V13.9852H66.2415V11.232H74.7727C74.7663 10.7279 74.6467 10.2787 74.414 9.88443C74.1814 9.49018 73.8614 9.18319 73.4543 8.96344C73.0536 8.73723 72.5914 8.62413 72.0679 8.62413C71.538 8.62413 71.0629 8.7437 70.6428 8.98283C70.2227 9.2155 69.8899 9.53542 69.6443 9.9426C69.3987 10.3433 69.2694 10.799 69.2565 11.3095V14.1113C69.2565 14.7188 69.3761 15.252 69.6152 15.7109C69.8543 16.1633 70.1936 16.5155 70.6331 16.7676C71.0726 17.0197 71.5961 17.1457 72.2037 17.1457C72.6238 17.1457 73.0051 17.0875 73.3476 16.9712C73.6902 16.8548 73.9842 16.6836 74.2298 16.4574C74.4754 16.2312 74.6596 15.9532 74.7824 15.6236L79.1353 15.7497C78.9543 16.7256 78.5568 17.5755 77.9429 18.2993C77.3353 19.0167 76.5371 19.5758 75.5483 19.9765C74.5594 20.3708 73.4155 20.5679 72.1164 20.5679Z"
-        fill="white"
-      />
-      <path d="M86.3457 0.43219V20.2867H81.605V0.43219H86.3457Z" fill="white" />
-      <path
-        d="M97.7683 5.39583V8.88588H88.3742V5.39583H97.7683ZM90.3422 1.82821H95.0829V15.6042C95.0829 15.8951 95.1281 16.131 95.2186 16.312C95.3156 16.4865 95.4545 16.6125 95.6355 16.69C95.8165 16.7611 96.033 16.7967 96.285 16.7967C96.466 16.7967 96.6567 16.7805 96.857 16.7482C97.0638 16.7094 97.219 16.6771 97.3224 16.6513L98.0398 20.0735C97.8136 20.1381 97.4936 20.2189 97.08 20.3158C96.6728 20.4128 96.1849 20.4742 95.6161 20.5C94.5045 20.5517 93.5512 20.4225 92.7562 20.1122C91.9677 19.7955 91.3634 19.3044 90.9433 18.6387C90.5297 17.973 90.3293 17.136 90.3422 16.1278V1.82821Z"
-        fill="white"
-      />
-      <rect
-        x="17.85"
-        y="10.4952"
-        width="19.95"
-        height="4.725"
-        transform="rotate(-15 17.85 10.4952)"
-        fill="white"
-      />
-    </svg>
-  );
-};
