+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "velt-website-chakra",
+  "name": "@velt/website",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -23,14 +23,21 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
+    "@bjerk/eslint-config": "^3.0.1",
     "@chakra-ui/cli": "^2.2.1",
     "@cobraz/prettier": "^2.0.1",
-    "eslint": "8.29.0",
-    "eslint-config-next": "13.0.6",
-    "prettier": "3.0.0-alpha.4",
     "@types/node": "18.11.12",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.9",
+    "@typescript-eslint/eslint-plugin": "5.49.1-alpha.0",
+    "eslint": "8.29.0",
+    "eslint-config-next": "13.0.6",
+    "eslint-config-prettier": "8.6.0",
+    "eslint-plugin-eslint-comments": "3.2.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-unicorn": "44.0.2",
+    "prettier": "^2.8.3",
     "typescript": "4.9.4"
   }
 }
