--- pages/_app.tsx
@@ -1,21 +1,0 @@
-import type { AppProps } from 'next/app';
-import { ChakraProvider } from '@chakra-ui/react';
-import { extendTheme } from '@chakra-ui/react';
-import theme from '../theme';
-
-export interface AppRenderProps {
-  pageProps: object;
-  err?: Error;
-  Component: NextComponentType<NextPageContext, AppRenderProps, object>;
-  router: NextRouter;
-}
-import type { NextComponentType, NextPageContext } from 'next';
-import type { NextRouter } from 'next/router';
-
-export default function App({ Component, pageProps }: AppRenderProps) {
-  return (
-    <ChakraProvider theme={theme}>
-      <Component {...pageProps} />
-    </ChakraProvider>
-  );
-}
