+++ src/components/header.tsx
@@ -1,0 +1,10 @@
+import { Flex } from '@chakra-ui/react';
+import { Logo } from './logo';
+
+export const Header: React.FC = () => {
+  return (
+    <Flex>
+      <Logo />
+    </Flex>
+  );
+};
