+++ src/pages/_app.tsx
@@ -1,0 +1,13 @@
+import { ChakraProvider } from '@chakra-ui/react';
+import type { AppProps } from 'next/app';
+import { theme } from '../theme';
+
+export const app: React.FC<AppProps> = ({ Component, pageProps }) => {
+  return (
+    <ChakraProvider theme={theme}>
+      <Component {...pageProps} />
+    </ChakraProvider>
+  );
+};
+
+export default app;
