+++ src/pages/index.tsx
@@ -1,0 +1,7 @@
+import { NextPage } from 'next';
+import React from 'react';
+import { Layout } from '../components/layout';
+
+const Home: NextPage = () => <Layout></Layout>;
+
+export default Home;
